feat: Use FullyConnected architecture (#143)
* feat: Use FullyConnected architecture

Braket has a field to indicate that a device is fully connected that
maps well onto the pytket FullyConnected architecture. This commit
changes the utility function that converts between the two models to
make use of this. This will change the BackendInfo returned by
simulators or devices with full connectivity.